Configure pci feature source
Configure the pci feature source to include all PCI ID fields in the
name of the published node labels. This is done by directly editing the
example conf (which gets deployed inside the container image that is
built). This is handy when using custom images, although the preferred
way would be to use ConfigMap and VolumeMount.